+++ src/build_number.h
@@ -1,1 +1,1 @@
-#define BUILD 58
+#define BUILD 59