+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.58
 - fixed a problem with linking notes with special characters in their name (everything but charactes and numbers)
+- html entities from the titles of inserted links are now decoded
 
 ## 0.57
 - `Ctrl + Click` in edit mode now works on the whole portion of links like `[QOwnNotes - cross-platform open source notepad](http://www.qownnotes.org/)`, not just on the URL alone